refactor(cli): separate JS command parser modules
