+++ packages/cli/binding/src/js_command_args/commands/mod.rs
@@ -1,0 +1,6 @@
+mod common;
+mod config;
+mod create;
+mod hooks;
+mod migrate;
+mod staged;
