+++ packages/cli/binding/src/js_command_args/mod.rs
@@ -1,6 +1,2 @@
-mod config;
-mod create;
-mod hooks;
-mod migrate;
-mod parse;
-mod staged;
+mod commands;
+mod parser;
