+++ packages/cli/binding/src/js_command_args/parser.rs
@@ -1,7 +1,6 @@
 use std::iter;
 
 use clap::{Arg, ArgAction, Args, Command, FromArgMatches, error::ErrorKind};
-use napi::bindgen_prelude::{Either, Either3};
 use napi_derive::napi;
 
 #[napi(object, object_from_js = false)]
@@ -10,7 +9,7 @@ pub struct CliParseError {
     pub message: String,
 }
 
-pub(super) enum ParseResult<T> {
+pub(in crate::js_command_args) enum ParseResult<T> {
     Ok(T),
     Help(Box<Command>),
     Error(CliParseError),
@@ -18,7 +17,7 @@ pub(super) enum ParseResult<T> {
 
 #[cfg(test)]
 impl<T> ParseResult<T> {
-    pub(super) fn expect_ok(self) -> T {
+    pub(in crate::js_command_args) fn expect_ok(self) -> T {
         match self {
             Self::Ok(value) => value,
             Self::Help(_) => panic!("The parser returned help."),
@@ -26,7 +25,7 @@ impl<T> ParseResult<T> {
         }
     }
 
-    pub(super) fn expect_error(self) -> CliParseError {
+    pub(in crate::js_command_args) fn expect_error(self) -> CliParseError {
         match self {
             Self::Error(error) => error,
             Self::Ok(_) => panic!("The parser returned arguments."),
@@ -35,7 +34,10 @@ impl<T> ParseResult<T> {
     }
 }
 
-pub(super) fn parse_args<T>(mut command: Command, argv: Vec<String>) -> ParseResult<T>
+pub(in crate::js_command_args) fn parse_args<T>(
+    mut command: Command,
+    argv: Vec<String>,
+) -> ParseResult<T>
 where
     T: Args + FromArgMatches,
 {
@@ -61,34 +63,8 @@ where
     }
 }
 
-pub(super) fn help_arg() -> Arg {
+pub(in crate::js_command_args) fn help_arg() -> Arg {
     Arg::new("help").short('h').long("help").action(ArgAction::Help).help("Show this help message")
-}
-
-pub(super) fn agent_option(
-    agent: Vec<String>,
-    no_agent: bool,
-) -> Option<Either3<bool, String, Vec<String>>> {
-    if no_agent {
-        Some(Either3::A(false))
-    } else if agent.len() == 1 {
-        agent.into_iter().next().map(Either3::B)
-    } else if agent.is_empty() {
-        None
-    } else {
-        Some(Either3::C(agent))
-    }
-}
-
-pub(super) fn editor_option(
-    mut editor: Vec<String>,
-    no_editor: bool,
-) -> Option<Either<bool, String>> {
-    if no_editor { Some(Either::A(false)) } else { editor.pop().map(Either::B) }
-}
-
-pub(super) fn boolean_option(enabled: bool, disabled: bool) -> Option<bool> {
-    if disabled { Some(false) } else { enabled.then_some(true) }
 }
 
 fn error_kind_name(kind: ErrorKind) -> &'static str {
